% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dni_wolne.R
\docType{data}
\name{dni_wolne}
\alias{dni_wolne}
\title{Ustawowo wolne dni od pracy w Polsce}
\format{data.frame}
\description{
dni_wolne Zbior danych dni_wolne zawiera ustawowo wolne dni od pracy w Polsce w latach 2001-2025
}

