% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wielkanoc.R
\name{is_wielkanoc}
\alias{is_boze_cialo}
\alias{is_poniedzialek_wielkanocny}
\alias{is_wielkanoc}
\alias{is_zielone_swiatki}
\title{Funkcje sprawdzajace czy data jest swietem ruchomym}
\usage{
is_wielkanoc(data)

is_poniedzialek_wielkanocny(data)

is_zielone_swiatki(data)

is_boze_cialo(data)
}
\arguments{
\item{data}{obiekt klasy Date}
}
\value{
Wektor dat
}
\description{
Funkcje pozwalaja na sprawdzenie czy zadany termin jest data wielkanocy, poniedzialku wielkanocnego, zielonych swiatek lub bozego ciala
}
\examples{
data <- as.Date("2016-03-27")
is_wielkanoc(data)
is_poniedzialek_wielkanocny(data)
is_zielone_swiatki(data)
is_boze_cialo(data)
}

