% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobotaniedziela.R
\name{sobotaniedziela}
\alias{is_niedziela}
\alias{is_sobota}
\alias{sobotaniedziela}
\title{Funkcje sprawdzajace czy dana data jest sobota lub niedziela}
\usage{
is_sobota(data)

is_niedziela(data)
}
\arguments{
\item{data}{obiekt klasy Date}
}
\value{
TRUE/FALSE
}
\description{
Funkcje sluzace do sprawdzenia czy podana data jest sobota lub niedziela
}
\examples{
data1 <- as.Date("2012-01-01")
data2 <- as.Date("2012-01-03")
data3 <- as.Date("2012-01-07")
is_sobota(data1)
is_sobota(data2)
is_sobota(data3)
is_niedziela(data1)
is_niedziela(data2)
is_niedziela(data3)
}

