% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wielkanoc.R
\name{wielkanoc}
\alias{boze_cialo}
\alias{poniedzialek_wielkanocny}
\alias{wielkanoc}
\alias{zielone_swiatki}
\title{Funkcje liczace daty ruchomych swiat}
\usage{
wielkanoc(rok)

poniedzialek_wielkanocny(rok)

zielone_swiatki(rok)

boze_cialo(rok)
}
\arguments{
\item{rok}{wektor numeryczny zawierajacy wybrany rok/lata}
}
\value{
Wektor dat
}
\description{
Funkcje pozwalaja na wyliczenie daty wielkanocy, poniedzialku wielkanocnego, zielonych swiatek i bozego ciala
}
\examples{
wielkanoc(2010)
poniedzialek_wielkanocny(2012)
zielone_swiatki(c(2016, 2010))
boze_cialo(2017)
}

